refactor(client): P7 — ToolLoopRunner + CompletionClassifier (#139)
* refactor(client): P7(1/2) — CompletionClassifier to completion.py

Verbatim moves: classifier prompt (module constant, bot keeps the
_CLASSIFIER_SYSTEM_PROMPT class-attr alias tests read), classify (was
_classify_completion; get_llm_client provider-backed), parse_response
(was _parse_classifier_response; bot keeps staticmethod delegate).
Reviewer boundary note honored: tool-loop policy machinery, not a
general service.

* refactor(client): P7(2/2) — chat tool loop to tool_loop.ToolLoopRunner

The 790-line _process_with_tools body moves verbatim into
ToolLoopRunner.run (host-based access, same pattern as the P5a dispatch
registry): iteration loop, context compression, guard cascade, stuck
tracking, classifier continuations, validation enforcement, parallel
execution + timeouts, audit/trajectory, vision injection, skill handoff,
/stop checkpoints. Guard detectors/messages import from response_guards
(the single source); truncate_tool_output from response_guards (identical
copy); client-module globals (_ALLOWED_WEBHOOK_IDS, fallback text, email
scrub) late-imported to avoid the module cycle — the webhook set is
rebound at startup and the call-time import reads the live binding.

ensure_failure_visible becomes a tool_loop module function; the bot keeps
a staticmethod alias (loop wrapper still calls it). Bot._process_with_
tools stays as the facade delegate (web/chat + tests call it).

Normalizations in the move (disclosed): _MAX_VALIDATION_RETRIES renamed
_max_validation_retries (function-local, N806); ruff format wrapped long
lines; three string literals manually wrapped. Orphaned client.py imports
removed (guard detectors, set_turn, _ERROR_RESULT_PREFIXES); the
SEND_MAX_RETRIES re-export gains an explicit noqa.

client.py: 3,850 -> ~3,150 lines. Characterization 127 green (zero test
changes), full suite 6040/4, tool_loop.py + completion.py ruff-clean.

NOTE per PR body: the promised decomposition of run() into named methods
rides with P8, where LoopPolicy parameterization forces exactly those
seams — one restructuring pass over these lines instead of two.

* docs: RFC R5 — record P7 move-only deviation (decomposition rides with P8)
